requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
requests
cryptography
python-multipart
